This is the only commit.
This is a fun project I did early in the Claim Academy Bootcamp. It has data that loads when the project is started. It has a console menu, and handles input with validation. Also allows the ability to paste the improperly formatted data the project required to be pasted in as a metric.  Fun with Java!